Faster srpm pairing with rpms
Instead of parsing huge number of RPMs to perform the calculation, use
the DNF repodata.

This actually also fixes an issue where we removed the source RPM for
each RPM, even when **not all** the binary RPMs from the given build
were subject of removal.  This use-case is now tested.

Fix: #7